+++ FoodDeliveryApp/src/main/java/com/hexaware/hotpot/controller/MenuItemsController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -30,5 +31,12 @@ public class MenuItemsController {
     @PreAuthorize("hasAuthority('customer')")
 	public List<MenuItems> searchMenuItems(@PathVariable String keyword) throws MenuItemNotFoundException{
 		return service.searchMenuItemsByKeyword(keyword);
+	}
+	
+	@DeleteMapping("/deleteMenu/{menuId}")
+    @PreAuthorize("hasAuthority('admin') or hasAuthority('manager')")
+	public String deleteMenu(@PathVariable long menuId) {
+		service.deleteMenu(menuId);
+		return "menu deleted successfully";
 	}
 }
