+++ FoodDeliveryApp/src/main/java/com/hexaware/hotpot/services/IMenuService.java
@@ -10,4 +10,6 @@ public interface IMenuService {
 	public List<MenuItems> getMenuByCategory(String category) throws MenuItemNotFoundException;
 
 	public List<MenuItems> searchMenuItemsByKeyword(String keyword) throws MenuItemNotFoundException;
+	
+	public void deleteMenu(long menuId);
 }
