+++ FoodDeliveryApp/src/main/resources/application.properties
@@ -5,7 +5,7 @@
 spring.jpa.hibernate.ddl-auto=update 
 spring.datasource.url=jdbc:mysql://localhost:3306/food_delivery_app
 spring.datasource.username=root
-spring.datasource.password=Sid@2002
+spring.datasource.password=Nipurna@123
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.show-sql=true
